scripts/requirements.txt
flask
flask-mysqldb
werkzeug
mysql-connector-python
python-dotenv
flask_sqlalchemy
qrcode[pil]
bakong-khqr
pyotp
flask-mail
pytesseract
Pillow
pdf2image
pyzbar
gunicorn
